+++ packages/kolme/tests/p2p-large.rs
@@ -169,13 +169,7 @@ async fn large_sync_inner(testtasks: TestTasks, (): ()) {
     assert_eq!(latest_block_height.next(), kolme1.read().get_next_height());
 
     // And now launch a gossip node for this Kolme
-    testtasks.try_spawn_persistent(
-        GossipBuilder::new()
-            .set_local_display_name("kolme1")
-            .build(kolme1)
-            .unwrap()
-            .run(),
-    );
+    let discovery = testtasks.launch_kademlia_discovery(kolme1, "kolme1");
 
     // We'll check at the end of the run to confirm that this never received the latest block.
     // First check that StateTransfer works
@@ -186,16 +180,16 @@ async fn large_sync_inner(testtasks: TestTasks, (): ()) {
     )
     .await
     .unwrap();
-    testtasks.try_spawn_persistent(
-        GossipBuilder::new()
-            .set_local_display_name("kolme_state_transfer")
-            .set_sync_mode(
+    testtasks.launch_kademlia_client_with(
+        kolme_state_transfer.clone(),
+        "kolme_state_transfer",
+        &discovery,
+        |builder| {
+            builder.set_sync_mode(
                 SyncMode::StateTransfer,
                 DataLoadValidation::ValidateDataLoads,
             )
-            .build(kolme_state_transfer.clone())
-            .unwrap()
-            .run(),
+        },
     );
 
     // Due to data size, it can take a bit to transfer the entire state